feat(welearn): rename welearn/query to welearn/fetch
